Custodian Balenciaga 2.0 Edition
This is essentially my first PR in which I have been helped by really great coders. Mostly Gundam on this.

This PR improves the ingame-presence of the Custodian Lore, and respectively removes the lore of the other downstream that had no relation to our current lore.

This PR also provides something the Custodians lacked: Loadout clothings and paves the way for the planned EOTP system that conveys their need to purify + in the future help other departments.