lots of work haha
upgrades the HP of several roaches, this is meant to let them withstand a hit without explode (as happens every time) at the same time rebalances some of the roach variants hp and resistances.

adds more details to the lost colony iron eagle overrun base and reduces the loot.

upgrades the sprite of the buckshot and adds only one, meant to reduce the lag of the shot and improve the sprite of it.

edits the recipes of the TT Nanomatter guns making them less hard to make but costing more materials and time.

edits the damage of the TT gun dealing less damage but with a bonus of burn damage, the bullet is halfway forged after all and does 14 damage = 8 burn and 8 brute, 10 ap.

edits the Iron Eagle Troopers helmets and uniforms giving them a face protector (as they should have had, i forgor) and also the terran federation banner on the shoulder (i forgor to add it in the first place)

removes temporairly the Mercenary Base from the crashsite (is gonna be re added soon with a new map and new mobs protecting it besides the Mercenaries)

rebalances the scrap sword, it was dealing an abusive ammount of damage, even better than a katana or TT diamond blade!